requirements.txt
loguru==0.7.0
PyYAML==6.0.1
thop==0.1.1.post2209072238
torch==1.11.0+cu113
